Extract duplicate test name detection to UniqueTestCaseNameResolver
- Also scope unique names to the test source
