+++ GherkinSpec.TestAdapter.UnitTests/MockTestCaseDiscoverySink.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.VisualStudio.TestPlatform.ObjectModel;
+using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
+using System.Collections.Generic;
+
+namespace GherkinSpec.TestAdapter.UnitTests
+{
+    class MockTestCaseDiscoverySink : ITestCaseDiscoverySink
+    {
+        private readonly List<TestCase> discoveredTests = new List<TestCase>();
+
+        public IReadOnlyList<TestCase> DiscoveredTests
+            => discoveredTests.AsReadOnly();
+
+        public void SendTestCase(TestCase discoveredTest)
+        {
+            discoveredTests.Add(discoveredTest);
+        }
+    }
+}
