+++ GherkinSpec.TestAdapter.UnitTests/NullMessageLogger.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
+
+namespace GherkinSpec.TestAdapter.UnitTests
+{
+    class NullMessageLogger : IMessageLogger
+    {
+        public void SendMessage(TestMessageLevel testMessageLevel, string message)
+        {
+        }
+    }
+}
